+++ src/main/java/com/cisowski/schoolmanagement/timetable/scheduleOccurrence/repository/ScheduleOccurrenceRepository.java
@@ -29,4 +29,8 @@ public interface ScheduleOccurrenceRepository extends JpaRepository<ScheduleOccu
     Set<ScheduleOccurrenceEntity> findByScheduleInAndOccurrenceDateTimeBetween(List<ScheduleEntity> schedules, LocalDateTime thresholdStartTime, LocalDateTime thresholdEndTime);
 
     List<ScheduleOccurrenceEntity> findByStatusInAndOccurrenceDateTimeBefore(List<OccurrenceStatus> statuses, LocalDateTime occurrenceExpirationTime);
+
+    boolean existsByScheduleAndStatusIn(ScheduleEntity schedule, List<OccurrenceStatus> statuses);
+
+    List<ScheduleOccurrenceEntity> findAllBySchedule(ScheduleEntity schedule);
 }