+++ src/main/java/com/cisowski/schoolmanagement/timetable/scheduleOccurrence/service/ScheduleOccurrenceService.java
@@ -17,4 +17,5 @@ public interface ScheduleOccurrenceService {
     ScheduleOccurrenceEntity fetchOccurrence(Long scheduleOccurrenceId);
     List<ScheduleOccurrenceEntity> findOccurrencesReadyToComplete(List<OccurrenceStatus> acceptableStatuses, Integer daysGap);
     void changeOccurrencesStatus(List<ScheduleOccurrenceEntity> occurrences, OccurrenceStatus targetStatus);
+    List<ScheduleOccurrenceEntity> fetchOccurrencesForSchedule(ScheduleEntity schedule);
 }